+++ tests/test_ut_content_parser_mkdn.py
@@ -571,6 +571,75 @@ class TestUtContentParserMkdn(object):  # pylint: disable=too-many-lines
         assert resource.uuid == 'f21c6318-8830-11e8-a114-2c4d54508088'
         assert resource.digest == '3306409c0901e27d754a3273a5964652e2e8ea80fe82f3aa80d1aef6e8ab8cef'
 
+    def test_parser_snippet_008(self):
+        """Test parsing snippet.
+
+        Test case verifies that the MKDN metadata is parsed correctly. In this
+        case there are new whitespaces at the end of metadata fields and the
+        ``source`` field is empty. The Tags must be read correctly here.
+        """
+
+        text = Const.NEWLINE.join((
+            '# Remove all exited containers and dangling images @docker',
+            '',
+            '> Remove all exited containers and dangling images. The command examples  ',
+            'first remove all exited containers and the all dangling images.',
+            '',
+            '> [1] https://docs.docker.com/engine/reference/commandline/images/  ',
+            '[2] https://docs.docker.com/engine/reference/commandline/rm/',
+            '',
+            '- Remove all exited containers',
+            '',
+            '    `$ docker rm $(docker ps --all -q -f status=exited)`',
+            '',
+            '- Remove all dangling images',
+            '',
+            '    `$ docker images -q --filter dangling=true | xargs docker rmi`',
+            '',
+            '## Meta',
+            '',
+            '> category  : snippet',
+            'created   : 2017-10-12T11:52:11.000001+00:00',
+            'digest    : 0a8b31f0ab442991e56dcaef1fc65aa6bff479c567e04dd7990948f201187c69',
+            'filename  : snippet.txt',
+            'languages : language',
+            'name      : example text',
+            'source    :',
+            'tags      : cleanup, container, docker, docker-ce, moby',
+            'updated   : 2017-10-12T11:52:11.000001+00:00',
+            'uuid      : f21c6318-8830-11e8-a114-2c4d54508088',
+            'versions  : git<=1.1.1,python>=2.7.0,python==3.7.0',
+            '',
+        ))
+        collection = Collection()
+        Parser(self.TIMESTAMP, text, collection).read_collection()
+        resource = next(collection.resources())
+        assert resource.category == Const.SNIPPET
+        assert resource.data == (
+            'docker rm $(docker ps --all -q -f status=exited)  #  Remove all exited containers',
+            'docker images -q --filter dangling=true | xargs docker rmi  #  Remove all dangling images'
+        )
+        assert resource.brief == 'Remove all exited containers and dangling images'
+        assert resource.description == (
+            'Remove all exited containers and dangling images. The command examples ' +
+            'first remove all exited containers and the all dangling images.'
+        )
+        assert resource.name == 'example text'
+        assert resource.groups == ('docker',)
+        assert resource.tags == ('cleanup', 'container', 'docker', 'docker-ce', 'moby')
+        assert resource.links == (
+            'https://docs.docker.com/engine/reference/commandline/images/',
+            'https://docs.docker.com/engine/reference/commandline/rm/'
+        )
+        assert resource.source == ''
+        assert resource.versions == ('git<=1.1.1', 'python==3.7.0', 'python>=2.7.0')
+        assert resource.languages == ('language',)
+        assert resource.filename == 'snippet.txt'
+        assert resource.created == '2017-10-12T11:52:11.000001+00:00'
+        assert resource.updated == '2017-10-12T11:52:11.000001+00:00'
+        assert resource.uuid == 'f21c6318-8830-11e8-a114-2c4d54508088'
+        assert resource.digest == '3306409c0901e27d754a3273a5964652e2e8ea80fe82f3aa80d1aef6e8ab8cef'
+
     def test_parser_solution_001(self):
         """Test parsing solution.