+++ CHANGELOG.rst
@@ -31,6 +31,7 @@ Bugfixes
 * Fix ``--sall`` option did not search all attributes.
 * Fix reading one line attributes like ``brief`` from multiple lines.
 * Fix terminal colors on Windows.
+* Fix parsing Markdown metadata from solution.
 * All commits in `0.11.0`_.
 
 Security